+++ cli/src/yagent/command_option.py
@@ -14,7 +14,6 @@ from yagent.commands.pdf.click import pdf_group
 from yagent.commands.dev.click import dev_group
 from yagent.commands.assoc import assoc_group, unassoc_group
 from yagent.commands.entity.click import entity_group
-from yagent.commands.tag.click import tag_group
 from yagent.commands.reminder.click import reminder_group
 from yagent.commands.routine.click import routine_group
 from yagent.commands.english.click import english_group
@@ -47,7 +46,6 @@ cli.add_command(image_group)
 cli.add_command(pdf_group)
 cli.add_command(dev_group)
 cli.add_command(entity_group)
-cli.add_command(tag_group)
 cli.add_command(assoc_group)
 cli.add_command(unassoc_group)
 cli.add_command(reminder_group)